fix:別ファイルにてUsersPropertyViewHolderを定義
